+++ common/header-didattica.php
@@ -19,6 +19,9 @@
 			<a href="<?php echo $__application_base_path; ?>/didattica/sportello.php" class="btn btn-default navbar-btn btn-orange4" role="button"><span class="glyphicon glyphicon-object-align-horizontal"></span>&ensp;Sportelli </a>
 			<a href="<?php echo $__application_base_path; ?>/didattica/studente.php" class="btn btn-default navbar-btn btn-lima4" role="button"><span class="glyphicon glyphicon-user"></span>&ensp;Studenti </a>
 		<?php endif; ?>
+		<?php if($__settings->config->corsiDiRecupero) : ?>
+			<a href="<?php echo $__application_base_path; ?>/docente/corsoDiRecuperoVoti.php" class="btn btn-default navbar-btn btn-lightblue4" role="button"><span class="glyphicon glyphicon-repeat"></span>&ensp;Corsi di Recupero </a>
+			<?php endif; ?>
 		</ul>
 
 		<ul class="nav navbar-nav navbar-right top-navbar-nav">