+++ docente/letteraCarenze.php
@@ -184,7 +184,14 @@ $superata = ($voto >= 6)? '<span class="c39" style="color:#08661a;">superata</sp
 			<tr class="c18">
 				<td class="c2" colspan="1" rowspan="1"><p class="c17">
 					</p></td>
-				<td class="c2" colspan="1" rowspan="1"><?php echo $__docente_nome . ' ' . $__docente_cognome; ?>
+				<td class="c2" colspan="1" rowspan="1">
+					<?php
+					if(getSettingsValue('config','corsiDiRecuperoFirmaDocente', true)) {
+						echo $__docente_nome . ' ' . $__docente_cognome;
+					} else {
+						echo ' ';
+					}
+					?>
 				</td>
 			</tr>
 		</tbody>